+++ vllm/model_executor/layers/quantization/compressed_tensors/schemes/__init__.py
@@ -2,7 +2,8 @@
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 
 from .compressed_tensors_scheme import CompressedTensorsScheme
-from .compressed_tensors_w4a4_nvfp4 import CompressedTensorsW4A4Fp4, CompressedTensorsW4A4MXFp4
+from .compressed_tensors_w4a4_nvfp4 import CompressedTensorsW4A4Fp4
+from .compressed_tensors_w4a4_mxfp4 import CompressedTensorsW4A4MXFp4
 from .compressed_tensors_w4a16_24 import (W4A16SPARSE24_SUPPORTED_BITS,
                                           CompressedTensorsW4A16Sparse24)
 from .compressed_tensors_w4a16_nvfp4 import CompressedTensorsW4A16Fp4
@@ -22,6 +23,6 @@ __all__ = [
     "WNA16_SUPPORTED_BITS", "W4A16SPARSE24_SUPPORTED_BITS",
     "CompressedTensors24", "CompressedTensorsW4A16Fp4",
     "CompressedTensorsW4A4Fp4",
-    "CompressedTensorsW4A4MXFp4"
+    "CompressedTensorsW4A4MXFp4",
     "CompressedTensorsW8A8MXFp8"
 ]
