+++ vllm/model_executor/layers/quantization/compressed_tensors/schemes/compressed_tensors_w4a4_nvfp4.py
@@ -158,120 +158,3 @@ class CompressedTensorsW4A4Fp4(CompressedTensorsScheme):
             out = out + bias
         return out.view(*output_shape)
 
-
-
-
-def run_mxfp4_emulations(
-        x,
-        weight,
-        weight_scale
-):
-    # TODO: (Yi)
-    pass
-
-class CompressedTensorsW4A4MXFp4(CompressedTensorsScheme):
-
-    def __init__(self):
-        self.group_size = 32
-
-    @classmethod
-    def get_min_capability(cls) -> int:
-        if envs.VLLM_USE_NVFP4_CT_EMULATIONS:
-            return 80
-        return 100
-
-
-    def create_weights(self, layer: torch.nn.Module,
-                       output_partition_sizes: list[int],
-                       input_size_per_partition: int,
-                       params_dtype: torch.dtype, weight_loader: Callable,
-                       **kwargs):
-        output_size_per_partition = sum(output_partition_sizes)
-        layer.logical_widths = output_partition_sizes
-        layer.input_size_per_partition = input_size_per_partition
-        layer.output_size_per_partition = output_size_per_partition
-
-        # Weight
-        weight = ModelWeightParameter(data=torch.empty(
-            sum(output_partition_sizes),
-            input_size_per_partition // 2,
-            dtype=torch.uint8),
-                                      input_dim=1,
-                                      output_dim=0,
-                                      weight_loader=weight_loader)
-        layer.register_parameter("weight_packed", weight)
-
-        # Global Weight Scale
-        # weight_global_scale = PerTensorScaleParameter(
-        #     data=torch.empty(len(output_partition_sizes), dtype=torch.float32),
-        #     weight_loader=weight_loader)
-        # layer.register_parameter("weight_global_scale", weight_global_scale)
-
-        # Per Group Weight Scale
-        weight_scale = GroupQuantScaleParameter(data=torch.empty(
-            sum(output_partition_sizes),
-            input_size_per_partition // self.group_size,
-            dtype=torch.uint8,
-        ),
-                                                input_dim=1,
-                                                output_dim=0,
-                                                weight_loader=weight_loader)
-
-        layer.register_parameter("weight_scale", weight_scale)
-
-        # input_global_scale = PerTensorScaleParameter(
-        #     data=torch.empty(len(output_partition_sizes), dtype=torch.float32),
-        #     weight_loader=weight_loader)
-        # layer.register_parameter("input_global_scale", input_global_scale)
-
-
-    def process_weights_after_loading(self, layer) -> None:
-        return 
-        global_input_scale = layer.input_global_scale.max().to(torch.float32)
-        layer.input_global_scale = Parameter(global_input_scale,
-                                             requires_grad=False)
-
-        layer.weight_global_scale = Parameter(
-            layer.weight_global_scale.max().to(torch.float32),
-            requires_grad=False)
-
-        swizzled_weight_scale = self.swizzle_blockscale(layer.weight_scale)
-        layer.weight_scale_swizzled = Parameter(swizzled_weight_scale,
-                                                requires_grad=False)
-
-        # required by cutlass kernel; need Parameter, not ModelWeightParameter
-        layer.weight = Parameter(layer.weight_packed.data, requires_grad=False)
-
-        layer.alpha = Parameter(layer.input_global_scale *
-                                layer.weight_global_scale,
-                                requires_grad=False)
-
-    def apply_weights(self,
-                      layer: torch.nn.Module,
-                      x: torch.Tensor,
-                      bias: Optional[torch.Tensor] = None) -> torch.Tensor:
-
-        if envs.VLLM_USE_NVFP4_CT_EMULATIONS:
-            from vllm.model_executor.layers.quantization.utils.mxfp4_utils import (
-                    run_mxfp4_emulations_no_swizzle,
-                )
-            out = run_mxfp4_emulations_no_swizzle(
-                x=x,
-                weight=layer.weight_packed,
-                weight_scale=layer.weight_scale)
-            if bias is not None:
-                out = out + bias
-            return out
-
-        # output_dtype = x.dtype
-        # output_shape = [x.shape[0], layer.weight.shape[0]]
-
-        # # quantize BF16 or FP16 to (FP4 and interleaved block scale)
-        # x_fp4, x_blockscale = scaled_fp4_quant(x, layer.input_global_scale)
-
-        # out = cutlass_scaled_fp4_mm(x_fp4, layer.weight, x_blockscale,
-        #                             layer.weight_scale_swizzled,
-        #                             1 / layer.alpha, output_dtype)
-        # if bias is not None:
-        #     out = out + bias
-        # return out.view(*output_shape)
