+++ vllm/model_executor/layers/quantization/compressed_tensors/utils.py
@@ -16,7 +16,7 @@ def is_activation_quantization_format(format: str) -> bool:
         CompressionFormat.int_quantized.value,
         CompressionFormat.float_quantized.value,
         CompressionFormat.nvfp4_pack_quantized.value,
-        CompressionFormat.mxfp4_pack_quantized.value,
+        CompressionFormat.mxfp4_pack_quantized.value
     ]
     return format in _ACTIVATION_QUANTIZATION_FORMATS
